+++ crates/sdk-core/src/worker/mod.rs
@@ -202,12 +202,10 @@ pub struct WorkerConfig {
     /// the options to fail.
     pub max_worker_activities_per_second: Option<f64>,
 
-    /// Maximum number of eager activities that can be running concurrently. When nonzero, eager
-    /// activity execution will not be requested if it would cause the number of running eager
-    /// activities to exceed this value. The default of zero means unlimited and therefore only
-    /// bound by the activity slot supplier.
-    #[builder(default = 0)]
-    pub max_concurrent_eager_activity_execution_size: usize,
+    /// Maximum number of activity slots that may be reserved for eager execution when completing
+    /// a workflow task. The default is 3. Setting this to zero disables eager activity execution.
+    #[builder(default = 3)]
+    pub max_eager_activity_reservations_per_workflow_task: usize,
 
     /// If set false (default), shutdown will not finish until all pending evictions have been
     /// issued and replied to. If set true shutdown will be considered complete when the only
@@ -855,7 +853,6 @@ impl Worker {
                 metrics.clone(),
                 config.max_heartbeat_throttle_interval,
                 config.default_heartbeat_throttle_interval,
-                config.max_concurrent_eager_activity_execution_size,
                 config.graceful_shutdown_period,
                 config.local_timeout_buffer_for_activities,
             )